Big fix: use uint64_t instead of int

+++ examples/callback_primes_oop.cpp
@@ -1,14 +1,15 @@
 /// @example callback_primes_oop.cpp
-/// Objects derived from PrimeSieveCallback<T> can be
+/// Objects derived from PrimeSieveCallback<uint64_t> can be
 /// passed to the callback_primes() function.
 
 #include <primesieve.h>
+#include <stdint.h>
 #include <list>
 
-struct PrimeList : std::list<int>,
-                   PrimeSieveCallback<int>
+struct PrimeList : std::list<uint64_t>,
+                   PrimeSieveCallback<uint64_t>
 {
-  void callback(int prime) {
+  void callback(uint64_t prime) {
     this->push_back(prime);
   }
 };
